fix: update all SurrealStorage methods to use connection field
Migrated all 44 methods in SurrealStorage implementation to use the new
Arc<RwLock<ConnectionState>> pattern instead of direct self.db access.

Changes:
- Added connection extraction pattern to start of every method
- Pattern clones the Surreal handle (cheap, Arc internally)
- Bails with clear errors on Reconnecting/Failed states
- Fixed test type mismatch (token_count u32 vs usize)

All methods now follow the pattern:
  let db = {
    let state = self.connection.read().unwrap();
    match &*state {
      ConnectionState::Connected(db) => db.clone(),
      ConnectionState::Reconnecting => anyhow::bail!("Connection is reconnecting"),
      ConnectionState::Failed(msg) => anyhow::bail!("Connection failed: {}", msg),
    }
  };

Updated methods across all trait groups:
- Knowledge Graph (10 methods)
- Scoped Memory (10 methods)
- TaskStreams (6 methods)
- Hybrid Search (1 method)
- Mindmaps (5 methods)
- Graph-RAG (5 methods)
- Temporal History (2 methods)
- Helper methods (5 methods)

Build: cargo build -p surreal-memory succeeds
Tests: 9/14 integration tests pass (4 ignored for server-mode, 1 RocksDB lock race)

Completes Task 3 scope gap from reconnection-logic-implementation plan.

+++ crates/surreal-memory/tests/integration_test.rs
@@ -258,7 +258,7 @@ async fn test_task_stream_lifecycle() {
     let profile = profile_for("default");
     for idx in 0..4 {
         let mut mem = memory(&format!("Long step {}", idx + 2), "u1");
-        mem.token_count = Some((profile.summarization_threshold() / 4 + 1) as usize);
+        mem.token_count = Some((profile.summarization_threshold() / 4 + 1) as u32);
         s.add_to_task_stream("my-task", mem)
             .await
             .expect("seed summarization");
